refactor test_openapi.py
