+++ tests/data/expected/parser/openapi/openapi_parser_parse/with_import.py
@@ -1,0 +1,55 @@
+from pydantic import AnyUrl, BaseModel
+from typing import List, Optional
+from __future__ import annotations
+
+
+class Pet(BaseModel):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Pets(BaseModel):
+    __root__: List[Pet]
+
+
+class User(BaseModel):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Users(BaseModel):
+    __root__: List[User]
+
+
+class Id(BaseModel):
+    __root__: str
+
+
+class Rules(BaseModel):
+    __root__: List[str]
+
+
+class Error(BaseModel):
+    code: int
+    message: str
+
+
+class Api(BaseModel):
+    apiKey: Optional[str] = None
+    apiVersionNumber: Optional[str] = None
+    apiUrl: Optional[AnyUrl] = None
+    apiDocumentationUrl: Optional[AnyUrl] = None
+
+
+class Apis(BaseModel):
+    __root__: List[Api]
+
+
+class Event(BaseModel):
+    name: Optional[str] = None
+
+
+class Result(BaseModel):
+    event: Optional[Event] = None
