+++ tests/data/expected/parser/openapi/openapi_parser_parse_additional_properties/format.py
