+++ tests/data/expected/parser/openapi/openapi_parser_parse_additional_properties/with_import_format.py
