+++ tests/data/expected/parser/openapi/openapi_parser_parse_alias/fo_o/ba_r.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+from pydantic import BaseModel, Field
+
+
+class ThinG(BaseModel):
+    attribute_s: Optional[Dict[str, Any]] = Field(None, alias='attribute-s')
+
+
+class ThanG(BaseModel):
+    attributes: Optional[List[Dict[str, Any]]] = None
+
+
+class ClonE(ThinG):
+    pass
