+++ tests/data/expected/parser/openapi/openapi_parser_parse_array_enum/with_import_format.py
@@ -1,0 +1,20 @@
+from __future__ import annotations
+
+from enum import Enum
+from typing import List
+
+from pydantic import BaseModel
+
+
+class Type1Enum(Enum):
+    enumOne = 'enumOne'
+    enumTwo = 'enumTwo'
+
+
+class Type1(BaseModel):
+    __root__: List[Type1Enum]
+
+
+class Type2(Enum):
+    enumFour = 'enumFour'
+    enumFive = 'enumFive'
