+++ tests/data/expected/parser/openapi/openapi_parser_parse_enum_models/output_py37.py
