+++ tests/data/expected/parser/openapi/openapi_parser_parse_modular/collections.py
