+++ tests/data/expected/parser/openapi/openapi_parser_parse_nested_anyof/output.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from typing import List, Optional, Union
+
+from pydantic import BaseModel
+
+
+class Type1(BaseModel):
+    prop: Optional[str] = None
+
+
+class Type2(BaseModel):
+    prop: Optional[str] = None
+
+
+class Container(BaseModel):
+    contents: List[Union[Type1, Type2]]
