+++ tests/data/expected/parser/openapi/openapi_parser_parse_resolved_models/output.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+from typing import List, Optional
+
+from pydantic import BaseModel
+
+
+class Pet(BaseModel):
+    id: int
+    name: str
+    tag: Optional[str] = None
+
+
+class Pets(BaseModel):
+    __root__: List[Pet]
+
+
+class Error(BaseModel):
+    code: int
+    message: str
+
+
+class Resolved(BaseModel):
+    resolved: Optional[List[str]] = None
